signal choice
