+++ boards/DIM/Inc/Io/Io_RotarySwitch.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <stdint.h>
-#include "App_DriveMode.h"
+#include "App_RotarySwitch.h"
 
 void Io_RotarySwitch_ExtiCallback(uint16_t GPIO_Pin);
 enum DriveMode Io_RotarySwitch_GetDriveMode(void);
