+++ boards/DIM/Src/Io/Io_GpioExtiCallback.c
@@ -1,4 +1,4 @@
-#include "Io_DriveMode.h"
+#include "Io_RotarySwitch.h"
 
 void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin)
 {
