+++ scripts/codegen/CAN/App_CanMsgs.dbc
@@ -172,6 +172,9 @@ SG_ State : 0|8@1+ (1,0) [0|255] "" LOGGER
 BO_ 504 DIM_WATCHDOG_TIMEOUT: 8 DIM
 SG_ TASK_NAME : 0|64@1+ (1,0) [0|0] "" LOGGER
 
+BO_ 506 DIM_ROTARY_SWITCH: 1 DIM
+SG_ Drive_Mode : 0|3@1+ (1,0) [0|0] "" DCM
+
 BA_DEF_  "BusType" STRING ;
 BA_DEF_ BO_  "GenMsgCycleTime" INT 0 65535;
 BA_DEF_ SG_  "GenSigStartValue" INT 0 2147483647;
@@ -211,6 +214,7 @@ BA_ "GenMsgCycleTime" BO_ 413 10;
 BA_ "GenMsgCycleTime" BO_ 500 100;
 BA_ "GenMsgCycleTime" BO_ 501 5000;
 BA_ "GenMsgCycleTime" BO_ 503 10;
+BA_ "GenMsgCycleTime" BO_ 506 10;
 
 BA_ "GenSigStartValue" SG_ 2  tx_overflow_count 0;
 BA_ "GenSigStartValue" SG_ 2  rx_overflow_count 0;
@@ -239,3 +243,4 @@ VAL_ 205 State 0 "INIT" 1 "DRIVE" 2 "FAULT";
 VAL_ 305 State 0 "AIR_OPEN" 1 "AIR_CLOSED";
 VAL_ 413 State 0 "INIT" 1 "AIR_OPEN" 2 "AIR_CLOSED";
 VAL_ 503 State 0 "DRIVE";
+VAL_ 506 Drive_Mode 0 "DRIVE_MODE_1" 1 "DRIVE_MODE_2" 2 "DRIVE_MODE_3" 3 "DRIVE_MODE_4" 4 "DRIVE_MODE_5" 5 "DRIVE_MODE_INVALID";
